Moves client parts to a separate directory
